+++ v2/pkg/engine/plan/datasource_filter_collect_nodes_visitor.go
@@ -18,6 +18,8 @@ type nodesCollector struct {
 	nodes       *NodeSuggestions
 	report      *operationreport.Report
 	keys        []DSKeyInfo
+
+	maxConcurrency uint
 }
 
 type DSKeyInfo struct {
@@ -50,6 +52,12 @@ func (c *nodesCollector) collectNodes() {
 	visitors := make([]*collectNodesVisitor, len(c.dataSources))
 	reports := make([]*operationreport.Report, len(c.dataSources))
 
+	// Create a semaphore if concurrency is limited
+	var sem chan struct{}
+	if c.maxConcurrency > 0 {
+		sem = make(chan struct{}, c.maxConcurrency)
+	}
+
 	for i, dataSource := range c.dataSources {
 		walker := astvisitor.WalkerFromPool()
 		visitor := &collectNodesVisitor{
@@ -66,8 +74,17 @@ func (c *nodesCollector) collectNodes() {
 		visitors[i] = visitor
 		report := operationreport.Report{}
 		reports[i] = &report
+
+		if sem != nil {
+			sem <- struct{}{}
+		}
 		go func(walker *astvisitor.Walker, report *operationreport.Report) {
 			defer wg.Done()
+			defer func() {
+				if sem != nil {
+					<-sem
+				}
+			}()
 			defer walker.Release()
 
 			defer func() {
@@ -78,7 +95,6 @@ func (c *nodesCollector) collectNodes() {
 			}()
 
 			walker.Walk(c.operation, c.definition, report)
-
 		}(walker, &report)
 	}
 	wg.Wait()